requirements-dev.txt
# Development-only dependencies
# For production dependencies, see requirements.txt

# Testing tools
covdefaults
coverage
pytest
pytest-env

# Development tools
pre-commit
black
flake8
mypy
isort

# The following are pinned versions for development
# These override the versions in requirements.txt when installed
pyyaml==5.4.1
